requirements.txt
flask
kafka-python
praw
nltk
textblob
